requirements.txt
streamlit
google-generativeai
python-dotenv
korean_lunar_calendar
requests
streamlit-lottie
